requirements.txt
autogen-agentchat
autogen-ext
autogen-ext[openai]
autogen-ext[azure]
autogen-ext[magentic-one]
playwright
uv
streamlit
python-dotenv